backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
chromadb==0.4.18
langchain>=0.1.0
langchain-community==0.0.38
python-dotenv==1.0.0
groq==0.4.1
pydantic==2.5.0
numpy==1.24.3
aiofiles==23.2.1
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
huggingface_hub<=0.19.4